finish without concatinating tmpfiles when p is 1

+++ src/sqd.coffee
@@ -71,20 +71,24 @@ main = (options)->
               finishProcesses++
               debug and console.error "%d process(es) finished: %dms", finishProcesses, new Date().getTime() - startTime
               if finishProcesses is nProcess
-                cat = cp.spawn "cat", tmpfiles
-                wstream = if output then fs.createWriteStream output, flags: "a", highWaterMark: 1024 * 1024 * 1024 -1 else process.stdout
+                if tmpfiles.length
+                  cat = cp.spawn "cat", tmpfiles
+                  wstream = if output then fs.createWriteStream output, flags: "a", highWaterMark: 1024 * 1024 * 1024 -1 else process.stdout
 
-                showStderr cat.stderr, "cat"
-                wstream.on "error", (e)-> error "outputStream"
-                cat.stdout.on "error", (e)-> error "cat.stdout"
-                cat.stdin.on "error", (e)-> error "cat.stdin"
+                  showStderr cat.stderr, "cat"
+                  wstream.on "error", (e)-> error "outputStream"
+                  cat.stdout.on "error", (e)-> error "cat.stdout"
+                  cat.stdin.on "error", (e)-> error "cat.stdin"
 
-                cat.stdout.pipe wstream
+                  cat.stdout.pipe wstream
 
-                wstream.on "close", ->
-                  unlinkCounter = 0
-                  cb = -> onClose() if ++unlinkCounter is nProcess
-                  fs.unlink tmpfile, cb for tmpfile in tmpfiles
+                  wstream.on "close", ->
+                    unlinkCounter = 0
+                    cb = -> onClose() if ++unlinkCounter is nProcess and typeof onClose is "function"
+                    fs.unlink tmpfile, cb for tmpfile in tmpfiles
+                else
+                  onClose() if typeof onClose is "function"
+
         )
 
 error = (streamName, end)->
